Adjust the tokenizer for one-line input
Instead of the line/column, only the "position" is tracked.
This allows matching tokens from the current position (using the
match() method of compiled regular expressions), so the source
string is no longer updated with each token.

Also, instead of SyntaxError (which is a bit tricky to raise correctly
from Python code), use a custom exception.